+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `assertion()`, `ReporterTAP.onAssert()`, and `testBehaviorExp` macro to handle simple expression evaluation (is true, is false, etc)
 - `Wire.resetMocks()` and documentation
 - `shiftIn()` and `shiftOut()`
+- `CIConfig.is_default` to detect when the default configuration is used
 
 ### Changed
 - Rubocop expected syntax downgraded from ruby 2.6 to 2.5